Tp 62 check all current orders on each pickup point and its details

+++ src/main/java/tqs/loadconnect/core_backend/repositories/OrderRepository.java
@@ -5,7 +5,7 @@ import tqs.loadconnect.core_backend.models.Order;
 
 import java.util.List;
 
-public interface OrderRepository extends JpaRepository<Order, Long> {
+public interface OrderRepository extends JpaRepository<Order, Integer> {
 
     List<Order> findByPickupPointId(Integer pickupPointId);
